Project Scaffolding
This commit sets out the basic structure for the SPhoG back-end. It sets
up a flask application, with mongoengine and APScheduler integration and
exposes a module to can be used to hydrate the cache with some initial
images. It also sets out linter rules and provides setup instructions.
The project is laid out according to the PyPA guidelines.